requirements.txt
python-telegram-bot==20.3
pyTelegramBotAPI==4.10.0
